remove keytar in favor of config; delete if found in keytar
