+++ src/commands/account/import.ts
@@ -3,16 +3,14 @@ import chalk from "chalk";
 import { ethers } from "ethers";
 import * as emoji from "node-emoji";
 import { get } from "../../lib/prompt";
-import { displayDescription } from "../../lib/utils";
-import { requireKeytar } from "../../lib/wallet";
+import { displayDescription, updateConfig } from "../../lib/utils";
 
 export default class AccountImport extends Command {
 	public static description = "import an ethereum account";
 
 	public async run() {
 		displayDescription(this, AccountImport.description);
-		let signerPrivateKey: string;
-		let keytar: any;
+		let key: string;
 		try {
 			const input: any = await get({
 				signerPrivateKey: {
@@ -21,24 +19,19 @@ export default class AccountImport extends Command {
 					hidden: true,
 				},
 			});
-			signerPrivateKey = input.signerPrivateKey;
+			key = input.key;
 		} catch (e) {
 			this.log("\n\nCancelled.\n");
 			process.exit(0);
 		}
-		try {
-			keytar = requireKeytar();
-			const wallet = new ethers.Wallet(signerPrivateKey);
-			await keytar.setPassword("airswap-cli", "private-key", signerPrivateKey);
-			this.log(
-				`\n${emoji.get("white_check_mark")} Set account to address ${chalk.bold(
-					wallet.address,
-				)}\n`,
-			);
-		} catch (e) {
-			this.log(
-				`${chalk.yellow("Error")} Cannot import account because dependencies are missing.\nIf you are on Linux, try installing libsecret-1-0 (Debian, Ubuntu etc.) or libsecret (RedHat, Fedora etc.) and then reinstalling AirSwap CLI.\n`,
-			);
-		}
+		const wallet = new ethers.Wallet(key);
+		await updateConfig(this, {
+			key,
+		});
+		this.log(
+			`\n${emoji.get("white_check_mark")} Set account to address ${chalk.bold(
+				wallet.address,
+			)}\n`,
+		);
 	}
 }
